front end js html and style files organized to separate folders in public.  Routes updated and all views load and display correctly
