+++ routes/html.js
@@ -30,15 +30,15 @@ const path = require("path");
 module.exports = function(router) {
 
     router.get("/stats", (req, res) => {
-        res.sendFile(path.join(__dirname, "../public/stats.html"))
+        res.sendFile(path.join(__dirname, "../public/views/stats.html"))
     });
     
     router.get("/", (req, res) => {
-        res.sendFile(path.join(__dirname, "../public/index.html"));
+        res.sendFile(path.join(__dirname, "../public/views/index.html"));
     });
     
     router.get("/exercise", (req, res) => {
-        res.sendFile(path.join(__dirname, "../public/exercise.html"));
+        res.sendFile(path.join(__dirname, "../public/views/exercise.html"));
     });
 
 }
