+++ routes/index.js
@@ -11,8 +11,8 @@ router.use('/api', apiRoutes);
 //HTML Routes
 // router.use('/html', htmlRoutes);
 
-// router.use(function(req, res){
-//     res.sendFile(path.join(__dirname, '../public/index.html'))
-// })
+router.use(function(req, res){
+    res.sendFile(path.join(__dirname, '../public/views/index.html'))
+})
 
 module.exports = router;
